Printer_Control_App/requirements.txt
appdirs==1.4.4
cairocffi==1.7.0
cairosvg==2.5.2
cffi==1.16.0
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.53.1
importlib-resources==6.4.0
kiwisolver==1.4.5
lxml==5.2.2
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.1
pandas==2.2.2
pillow==10.3.0
printrun==2.0.0rc7
psutil==5.9.8
pycparser==2.22
pyglet==1.5.29
pyparsing==3.1.2
pyqt5
pyqt5-qt5
pyqt5-sip
pyqtgraph
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
tinycss2==1.3.0
typing-extensions==4.11.0
tzdata==2024.1
webencodings==0.5.1
wxpython==4.2.1
xlsxwriter==3.2.0
xlwt==1.3.0
zipp==3.19.2
PyMuPDF